Add mongo command to deploy.sh script
Closes #69

+++ deploy.sh
@@ -1,6 +1,9 @@
 #!/bin/bash
 
-if [[ $# != 1 || ( $1 != "dev" && $1 != "prod" ) ]] ; then
+if [[ $# != 1 || (
+  $1 != "dev" &&
+  $1 != "prod" &&
+  $1 != "mongo" ) ]] ; then
   echo "Please enter a single argument to specify prod or dev"
   exit 0
 fi
@@ -19,11 +22,13 @@ if [[ $1 = "prod" ]] ; then
   sleep 5s
   docker cp secret.js cpps_app_1:/home/src
   docker exec -itd cpps_app_1 gulp
-else
+elif [[ $1 = "dev" ]] ; then
   docker-compose down
   docker-compose build
   docker-compose up &
   sleep 5s
   docker cp secret.js cpps_app_1:/home/src
   docker exec -it cpps_app_1 /bin/bash -c "cd /root/src && npm install && gulp"
+elif [[ $1 = "mongo" ]] ; then
+  docker exec -it cpps_db_1 mongo
 fi
